When main window is closed by user clicking on the 'x' icon, the user is alerted to modified files and allowed to save or not save them or cancel the close.
The tab for each modified file is selected so that file is displayed when
the user is queried.

When a new file is loaded and a tab created, that tab is now selected so
the new file is displayed.

In editor frame, "editorRig.loadFile(pFullpath);" is now used instead of
retrieving the new editorRig using getComponentAt. Both do the same thing
but the old way is illogical.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>